# docs/helixnet-openapi.json
{"openapi":"3.1.0","info":{"title":"HelixNet Core API","description":"HelixNet FastAPI backend secured via Keycloak OpenID Connect","version":"2.0.1"},"paths":{"/protected":{"get":{"tags":["🔐 Security"],"summary":"Protected Route","description":"Protected route for verifying Keycloak authentication.","operationId":"protected_route_protected_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2":[]}]}},"/api/v2/auth/token":{"post":{"tags":["🔑 Authentication","Authentication"],"summary":"Authenticate user and return JWT access and refresh tokens","description":"Authenticates the user against the local DB and issues JWT access + refresh tokens.","operationId":"login_for_access_token_api_v2_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_v2_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v2/auth/refresh":{"post":{"tags":["🔑 Authentication","Authentication"],"summary":"Refresh access token using a valid refresh token","description":"Validates and rotates a refresh token to issue a new access/refresh pair.","operationId":"refresh_token_api_v2_auth_refresh_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_refresh_token_api_v2_auth_refresh_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v2/users/register":{"post":{"tags":["👤 Users","Users"],"summary":"Register User","operationId":"register_user_api_v2_users_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v2/users/me":{"get":{"tags":["👤 Users","Users"],"summary":"🔒 Get current user profile","operationId":"read_current_user_api_v2_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["👤 Users","Users"],"summary":"🔒 Update current user profile","operationId":"update_current_user_api_v2_users_me_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v2/users/{user_id}":{"get":{"tags":["👤 Users","Users"],"summary":"🔒 Get user by ID","operationId":"read_user_api_v2_users__user_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["👤 Users","Users"],"summary":"🔒 Delete user (Admin/Owner)","operationId":"delete_user_api_v2_users__user_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"User Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v2/users/":{"get":{"tags":["👤 Users","Users"],"summary":"🔒 List all users (Admin only)","operationId":"list_users_api_v2_users__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserRead"},"title":"Response List Users Api V2 Users  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v2/jobs":{"get":{"tags":["⚙️ Jobs","Jobs"],"summary":"List Jobs","description":"List all jobs owned by the current authenticated user.","operationId":"list_jobs_api_v2_jobs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/JobRead"},"type":"array","title":"Response List Jobs Api V2 Jobs Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["⚙️ Jobs","Jobs"],"summary":"Create Job No Files","description":"Create a new background job.","operationId":"create_job_no_files_api_v2_jobs_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v2/jobs/{job_id}":{"get":{"tags":["⚙️ Jobs","Jobs"],"summary":"Get Job","description":"Retrieve a single job by ID.","operationId":"get_job_api_v2_jobs__job_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["⚙️ Jobs","Jobs"],"summary":"Delete Job","description":"Delete a job by ID.","operationId":"delete_job_api_v2_jobs__job_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Job Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v2/jobs/upload":{"post":{"tags":["⚙️ Jobs","Jobs"],"summary":"Create Job With Files","description":"Create a new job via multiple file uploads.","operationId":"create_job_with_files_api_v2_jobs_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_job_with_files_api_v2_jobs_upload_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobFileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/health/healthz":{"get":{"tags":["💓 Health"],"summary":"Healthz","operationId":"healthz_health_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/health":{"get":{"tags":["💓 Health"],"summary":"💖 Robust API Health Check","description":"Performs deep health checks on all critical, complex backend services: \n    **PostgreSQL** (DB) and **Celery** (Worker/Broker).\n    \n    A successful **HTTP 200 OK** response indicates all core dependencies are functional.\n    A **HTTP 503 Service Unavailable** response means at least one dependency is failing.","operationId":"health_check_health_health_get","responses":{"200":{"description":"Returns a JSON object detailing the status of each service.","content":{"application/json":{"schema":{"type":"object","title":"Response Health Check Health Health Get"}}}}}}},"/":{"get":{"tags":["🧭 Web UI"],"summary":"Dashboard","description":"Render main dashboard page.","operationId":"dashboard__get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/login":{"get":{"tags":["🔑 Web UI"],"summary":"Login Page","description":"Render login page.","operationId":"login_page_login_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/submit-form":{"get":{"tags":["📨 Web UI"],"summary":"Submit Form Page","description":"Render form submission page.","operationId":"submit_form_page_submit_form_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Body_create_job_with_files_api_v2_jobs_upload_post":{"properties":{"content_file":{"type":"string","format":"binary","title":"Content File"},"context_file":{"type":"string","format":"binary","title":"Context File"},"template_file":{"type":"string","format":"binary","title":"Template File"},"schema_file":{"type":"string","format":"binary","title":"Schema File"}},"type":"object","required":["content_file","context_file","template_file","schema_file"],"title":"Body_create_job_with_files_api_v2_jobs_upload_post"},"Body_login_for_access_token_api_v2_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_v2_auth_token_post"},"Body_refresh_token_api_v2_auth_refresh_post":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"Body_refresh_token_api_v2_auth_refresh_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobCreate":{"properties":{"title":{"type":"string","title":"Title"},"payload":{"type":"object","title":"Payload"}},"type":"object","required":["title"],"title":"JobCreate","description":"Represents data required to create a new job.","examples":[{"payload":{"complexity_level":9000,"priority":"HIGH","report_type":"quarterly_audit","target_user_ids":["9a326614-478f-4332-b65d-8824709cfa1e","b3f0c2e1-a7d9-4b8c-8c1d-6b0d9e5f4a7b"]},"title":"Triple Roundhouse Compliance Report"}]},"JobFileResponse":{"properties":{"job_id":{"type":"string","format":"uuid","title":"Job Id"},"status":{"type":"string","title":"Status"},"uploaded_files":{"additionalProperties":{"type":"string"},"type":"object","title":"Uploaded Files"}},"type":"object","required":["job_id","status","uploaded_files"],"title":"JobFileResponse","description":"The enriched response model for file uploads."},"JobRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"title":{"type":"string","title":"Title"},"status":{"type":"string","title":"Status"},"owner_email":{"type":"string","title":"Owner Email"}},"type":"object","required":["id","title","status","owner_email"],"title":"JobRead","description":"Represents a job record returned to the user."},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token","description":"The short-lived JWT used for API authorization."},"refresh_token":{"type":"string","title":"Refresh Token","description":"The long-lived JWT used to obtain new access tokens."},"token_type":{"type":"string","title":"Token Type","description":"The type of token.","default":"bearer"}},"type":"object","required":["access_token","refresh_token"],"title":"TokenResponse","description":"Schema for returning both access and refresh tokens after login or refresh."},"UserCreate":{"properties":{"username":{"type":"string","title":"Username","example":"chuck.norris"},"email":{"type":"string","format":"email","title":"Email","example":"chuck@norris.io"},"first_name":{"type":"string","title":"First Name","example":"Chuck"},"last_name":{"type":"string","title":"Last Name","example":"Norris"},"password":{"type":"string","title":"Password","example":"roundhouseKick42!"}},"type":"object","required":["username","email","first_name","last_name","password"],"title":"UserCreate"},"UserRead":{"properties":{"email":{"type":"string","format":"email","title":"Email","example":"jane.doe@helix.net"},"username":{"type":"string","minLength":3,"title":"Username","example":"janedoe"},"fullname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fullname","example":"Jane Doe"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number","description":"User's primary contact number.","example":"+1-555-123-4567"},"web_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Web Url","description":"Personal or professional web link.","example":"https://janedoe.com"},"id":{"type":"string","format":"uuid","title":"Id","description":"Unique identifier for the user."},"is_active":{"type":"boolean","title":"Is Active"},"is_superuser":{"type":"boolean","title":"Is Superuser","description":"True if the user has superuser/admin privileges.","default":false},"scopes":{"items":{"type":"string"},"type":"array","title":"Scopes","default":[]},"roles":{"items":{"type":"string"},"type":"array","title":"Roles","description":"List of user roles as strings (e.g., 'basic', 'admin').","default":[]},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["email","username","id","is_active","created_at"],"title":"UserRead","description":"The definitive read-only schema for a User object, containing all persisted data.\nPydantic V2 compliant and ready for ORM mapping."},"UserRoles":{"type":"string","enum":["basic","manager","admin","auditor","tester","developer","guest"],"title":"UserRoles","description":"Defines the available roles for users.\nStandardized, singular source of truth for all role checks."},"UserUpdate":{"properties":{"email":{"type":"string","format":"email","title":"Email","example":"jane.doe@helix.net"},"username":{"type":"string","minLength":3,"title":"Username","example":"janedoe"},"fullname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fullname","example":"Jane Doe"},"phone_number":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone Number","description":"User's primary contact number.","example":"+1-555-123-4567"},"web_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Web Url","description":"Personal or professional web link.","example":"https://janedoe.com"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"is_admin":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Admin","description":"Maps to the database superuser status."},"roles":{"anyOf":[{"items":{"$ref":"#/components/schemas/UserRoles"},"type":"array"},{"type":"null"}],"title":"Roles"},"password":{"anyOf":[{"type":"string","minLength":8},{"type":"null"}],"title":"Password","description":"Optional new password."}},"type":"object","required":["email","username"],"title":"UserUpdate","description":"Schema for updating an existing user's profile."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2":{"type":"oauth2","description":"OAuth2 Password flow via Keycloak","flows":{"password":{"scopes":{"openid":"Access Helix API"},"tokenUrl":"https://keycloak.helix.local/realms/master/protocol/openid-connect/token"}}},"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"http://keycloak:8080/realms/master/protocol/openid-connect/token"}}}}}}